requirements.txt
amqp==5.1.1
anyio==3.6.2
APScheduler==3.6.3
async-timeout==4.0.2
beautifulsoup4==4.12.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
emoji==2.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
kombu==5.2.4
prompt-toolkit==3.0.38
pyaes==1.6.1
pyasn1==0.4.8
pyTelegramBotAPI==4.10.0
python-telegram-bot==13.3
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
redis==4.5.4
reloading==1.1.2
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
Telethon==1.27.0
tornado==6.2
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
